Clear up directory structure
