--- underScore.egg-info/SOURCES.txt
@@ -1,16 +1,0 @@
-README.md
-setup.py
-tests/__init__.py
-tests/test_dummy_data.py
-tests/test_messages.py
-the_mines/__init__.py
-the_mines/get_data.py
-the_mines/get_html.py
-the_mines/scrape_html.py
-underScore.egg-info/PKG-INFO
-underScore.egg-info/SOURCES.txt
-underScore.egg-info/dependency_links.txt
-underScore.egg-info/top_level.txt
-utils/__init__.py
-utils/dummy_data.py
-utils/messages.py
