--- underScore.egg-info/top_level.txt
@@ -1,3 +1,0 @@
-tests
-the_mines
-utils
